% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read3.R
\name{convert_kdbx_headers}
\alias{convert_kdbx_headers}
\title{Convert raw KDBX header fields to usable R types}
\usage{
convert_kdbx_headers(headers)
}
\arguments{
\item{headers}{Named list of raw values}
}
\value{
Named list of converted values
}
\description{
Convert raw KDBX header fields to usable R types
}
