% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read3.R
\name{kdbx_decrypt}
\alias{kdbx_decrypt}
\title{Decrypt a KDBX v3.1 database}
\usage{
kdbx_decrypt(kdbx, password)
}
\arguments{
\item{kdbx}{Result from kdbx_read()}

\item{password}{Character string password}
}
\value{
Raw decrypted XML (without stream start bytes)
}
\description{
Decrypt a KDBX v3.1 database
}
