% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read3.R
\name{kdbx_read}
\alias{kdbx_read}
\title{Read KDBX 3.1 Header and Encrypted Payload}
\usage{
kdbx_read(file, convert = TRUE)
}
\arguments{
\item{file}{Path to .kdbx file}

\item{convert}{Logical. If TRUE, convert some header fields to usable types}
}
\value{
List: version, headers (named list), and encrypted (raw)
}
\description{
Read KDBX 3.1 Header and Encrypted Payload
}
