% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeepass-package.R
\docType{package}
\name{rkeepass-package}
\alias{rkeepass}
\alias{rkeepass-package}
\title{rkeepass: Read KeePass Database Files}
\description{
Read entries from KeePass '.kdbx' database files (versions 3.x and 4.x) using the 'keepass-rs' Rust library. Supports password and keyfile authentication.
}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

}
\keyword{internal}
